Move all script to component_completion
