+++ nf_core/components/components_utils.py
@@ -1,5 +1,7 @@
 import logging
 import re
+import sys
+
 from pathlib import Path
 from typing import Optional, Union
