--- tests/subworkflows/test_completion.py
@@ -1,83 +1,0 @@
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-from nf_core.subworkflows._completion import autocomplete_subworkflows
-
-
-class DummyParam:
-    # Minimal mock object for Click parameter (not used in the function)
-    pass
-
-
-class DummyCtx:
-    def __init__(self, obj=None):
-        self.obj = obj
-
-
-@patch("nf_core.subworkflows._completion.CompletionItem")
-@patch("nf_core.subworkflows._completion.SubworkflowList")
-def test_autocomplete_subworkflows_mocked(mock_subworkflows_list_class, mock_completion_item_class):
-    # Setup mock for module list
-    mock_instance = mock_subworkflows_list_class.return_value
-    mock_instance.modules_repo.get_avail_components.return_value = [
-        "vcf_gather_bcftools",
-        "fastq_align_star",
-        "utils_nextflow_pipeline",
-    ]
-
-    # Setup mock for CompletionItem
-    def mock_completion(value):
-        mock_item = MagicMock()
-        mock_item.value = value
-        return mock_item
-
-    mock_completion_item_class.side_effect = mock_completion
-
-    ctx = DummyCtx()
-    param = DummyParam()
-    completions = autocomplete_subworkflows(ctx, param, "utils")
-
-    values = [c.value for c in completions]
-    assert "utils_nextflow_pipeline" in values
-    assert "vcf_gather_bcftools" not in values
-
-
-@patch("nf_core.subworkflows._completion.SubworkflowList")
-def test_autocomplete_subworkflows_with_ctx_obj(mock_subworkflows_list_class):
-    # Setup mock return value
-    mock_instance = mock_subworkflows_list_class.return_value
-    mock_instance.modules_repo.get_avail_components.return_value = [
-        "vcf_gather_bcftools",
-        "fastq_align_star",
-        "utils_nextflow_pipeline",
-    ]
-
-    # Provide ctx.obj with custom values
-    ctx = DummyCtx(
-        obj={
-            "modules_repo_url": "https://custom.url/modules",
-            "modules_repo_branch": "custom-branch",
-            "modules_repo_no_pull": True,
-        }
-    )
-
-    param = DummyParam()
-    completions = autocomplete_subworkflows(ctx, param, "utils")
-
-    # Assertions
-    mock_subworkflows_list_class.assert_called_once_with(".", True, "https://custom.url/modules", "custom-branch", True)
-
-    values = [c.value for c in completions]
-    assert "utils_nextflow_pipeline" in values
-    assert "vcf_gather_bcftools" not in values
-
-
-def test_autocomplete_subworkflows_missing_argument():
-    ctx = DummyCtx()
-    param = DummyParam()
-
-    with pytest.raises(TypeError) as exc_info:
-        autocomplete_subworkflows(ctx, param)  # Missing 'incomplete' argument
-
-    assert "missing 1 required positional argument" in str(exc_info.value)
